pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mph-agent"
version = "0.1.0"
description = "一个智能 Agent，将自然语言描述的 2D/3D 建模需求（几何、材料、物理场）自动转换为 COMSOL Multiphysics 模型文件"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["comsol", "agent", "nlp", "simulation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "openai>=1.0.0",
    "requests>=2.28.0",
    "jpype1>=1.4.0",
    "rich>=13.7.0",
    "loguru>=0.7.2",
    "tqdm>=4.66.0",
    "setuptools>=61.0",
    "questionary>=1.10.0",
    "sqlite-vec>=0.1.0",
    "celery>=5.5.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
# 向量检索：安装后可使用 skills 的语义检索（否则仅关键词/trigger 匹配）
vec = [
    "sentence-transformers>=2.2.0",
]
# 记忆 Agent 后台：Celery + Redis，用于异步更新会话摘要
memory = [
    "celery>=5.3.0",
    "redis>=4.0.0",
]

# 不再设定 Python 包，仅保留桌面端与源码运行；使用 uv run python cli.py 启动

[tool.setuptools]
packages = ["agent", "schemas", "prompts", "prompts.planner", "prompts.react", "prompts.executor"]
py-modules = ["cli"]

[tool.setuptools.package-data]
agent = ["*.txt", "*.md"]
schemas = ["*.md"]
prompts = ["**/*.txt"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
